Rearrange files in extensions directories (#53)
* rearrange files

* add output folder content to gitignore

* minor changes (test changes, units, update default argument path)

* shorten output string - fix for saving files on windows

@p-wlich : this caused saving failures on windows because the string was too long
