+++ .gitignore
@@ -2,3 +2,4 @@ docs/build
 .ipynb_checkpoints
 docs/Manifest.toml
 .vscode
+output/*
