+++ src/components/extensions/AdaptationCosts_annualGrowth.jl
