+++ src/components/extensions/ClimateTemperature_annual.jl
