+++ src/components/extensions/EquityWeighting_annual.jl
