+++ src/components/extensions/EquityWeighting_annualGrowth.jl
