+++ src/components/extensions/EquityWeighting_growth.jl
