+++ src/components/extensions/MarketDamagesBurke_annualGrowth.jl
