+++ src/components/extensions/MarketDamages_annual.jl
