+++ src/components/extensions/MarketDamages_annualGrowth.jl
