+++ src/components/extensions/NonMarketDamages_annualGrowth.jl
