+++ src/components/extensions/TotalAdaptationCosts_annualGrowth.jl
