+++ src/extensions/climate_model_annual.jl
@@ -1,23 +1,23 @@
-include("utils/load_parameters.jl")
-include("components/RCPSSPScenario.jl")
-include("components/CO2emissions.jl")
-include("components/CO2cycle.jl")
-include("components/CO2forcing.jl")
-include("components/CH4emissions.jl")
-include("components/CH4cycle.jl")
-include("components/CH4forcing.jl")
-include("components/N2Oemissions.jl")
-include("components/N2Ocycle.jl")
-include("components/N2Oforcing.jl")
-include("components/LGemissions.jl")
-include("components/LGcycle.jl")
-include("components/LGforcing.jl")
-include("components/SulphateForcing.jl")
-include("components/TotalForcing.jl")
-include("components/ClimateTemperature_annual.jl")
-include("components/PermafrostSiBCASA_annual.jl")
-include("components/PermafrostJULES_annual.jl")
-include("components/PermafrostTotal_annual.jl")
+include("../utils/load_parameters.jl")
+include("../components/RCPSSPScenario.jl")
+include("../components/CO2emissions.jl")
+include("../components/CO2cycle.jl")
+include("../components/CO2forcing.jl")
+include("../components/CH4emissions.jl")
+include("../components/CH4cycle.jl")
+include("../components/CH4forcing.jl")
+include("../components/N2Oemissions.jl")
+include("../components/N2Ocycle.jl")
+include("../components/N2Oforcing.jl")
+include("../components/LGemissions.jl")
+include("../components/LGcycle.jl")
+include("../components/LGforcing.jl")
+include("../components/SulphateForcing.jl")
+include("../components/TotalForcing.jl")
+include("../components/extensions/ClimateTemperature_annual.jl")
+include("../components/extensions/PermafrostSiBCASA_annual.jl")
+include("../components/extensions/PermafrostJULES_annual.jl")
+include("../components/extensions/PermafrostTotal_annual.jl")
 
 function climatemodel(scenario::String, use_permafrost::Bool=true, use_seaice::Bool=true)
     m = Model()
