+++ src/extensions/compute_scc_annual.jl
