+++ src/extensions/compute_scc_annualGrowth.jl
