+++ src/extensions/compute_scc_growth.jl
