+++ src/extensions/main_model_ARvariability.jl
@@ -2,45 +2,45 @@ using Mimi
 
 export getpage
 
-include("utils/load_parameters.jl")
-include("utils/mctools.jl")
+include("../utils/load_parameters.jl")
+include("../utils/mctools.jl")
 
 include("mcs_ARvariability.jl")
 include("compute_scc_annual.jl")
 
-include("components/RCPSSPScenario.jl")
-include("components/CO2emissions.jl")
-include("components/CO2cycle.jl")
-include("components/CO2forcing.jl")
-include("components/CH4emissions.jl")
-include("components/CH4cycle.jl")
-include("components/CH4forcing.jl")
-include("components/N2Oemissions.jl")
-include("components/N2Ocycle.jl")
-include("components/N2Oforcing.jl")
-include("components/LGemissions.jl")
-include("components/LGcycle.jl")
-include("components/LGforcing.jl")
-include("components/SulphateForcing.jl")
-include("components/TotalForcing.jl")
-include("components/ClimateTemperature_ARvariability.jl")
-include("components/SeaLevelRise.jl")
-include("components/GDP.jl")
-include("components/MarketDamages_annual.jl")
-include("components/MarketDamagesBurke_annual.jl")
-include("components/NonMarketDamages_annual.jl")
-include("components/Discontinuity_annual.jl")
-include("components/AdaptationCosts.jl")
-include("components/SLRDamages.jl")
-include("components/AbatementCostParameters.jl")
-include("components/AbatementCosts.jl")
-include("components/TotalAbatementCosts.jl")
-include("components/TotalAdaptationCosts.jl")
-include("components/Population.jl")
-include("components/EquityWeighting_annual.jl")
-include("components/PermafrostSiBCASA.jl")
-include("components/PermafrostJULES.jl")
-include("components/PermafrostTotal.jl")
+include("../components/RCPSSPScenario.jl")
+include("../components/CO2emissions.jl")
+include("../components/CO2cycle.jl")
+include("../components/CO2forcing.jl")
+include("../components/CH4emissions.jl")
+include("../components/CH4cycle.jl")
+include("../components/CH4forcing.jl")
+include("../components/N2Oemissions.jl")
+include("../components/N2Ocycle.jl")
+include("../components/N2Oforcing.jl")
+include("../components/LGemissions.jl")
+include("../components/LGcycle.jl")
+include("../components/LGforcing.jl")
+include("../components/SulphateForcing.jl")
+include("../components/TotalForcing.jl")
+include("../components/extensions/ClimateTemperature_ARvariability.jl")
+include("../components/SeaLevelRise.jl")
+include("../components/GDP.jl")
+include("../components/extensions/MarketDamages_annual.jl")
+include("../components/extensions/MarketDamagesBurke_annual.jl")
+include("../components/extensions/NonMarketDamages_annual.jl")
+include("../components/extensions/Discontinuity_annual.jl")
+include("../components/AdaptationCosts.jl")
+include("../components/SLRDamages.jl")
+include("../components/AbatementCostParameters.jl")
+include("../components/AbatementCosts.jl")
+include("../components/TotalAbatementCosts.jl")
+include("../components/TotalAdaptationCosts.jl")
+include("../components/Population.jl")
+include("../components/extensions/EquityWeighting_annual.jl")
+include("../components/PermafrostSiBCASA.jl")
+include("../components/PermafrostJULES.jl")
+include("../components/PermafrostTotal.jl")
 
 function buildpage(m::Model, scenario::String, use_permafrost::Bool=true, use_seaice::Bool=true, use_page09damages::Bool=false)
