+++ src/extensions/mcs_growth.jl
@@ -2,7 +2,7 @@ using Distributions
 using DataFrames
 using Mimi
 
-include("utils/mctools.jl")
+include("../utils/mctools.jl")
 
 function getsim(ge_minimum::Union{Float64, Nothing} = nothing,
                 ge_maximum::Union{Float64, Nothing} = nothing,
@@ -257,7 +257,7 @@ function getsim(ge_minimum::Union{Float64, Nothing} = nothing,
 end
 
 #Reformat the RV results into the format used for testing
-function reformat_RV_outputs(samplesize::Int; output_path::String = joinpath(@__DIR__, "../output"))
+function reformat_RV_outputs(samplesize::Int; output_path::String = joinpath(@__DIR__, "../../output"))
 
     #create vectors to hold results of Monte Carlo runs
     td=zeros(samplesize);
@@ -297,28 +297,9 @@ function reformat_RV_outputs(samplesize::Int; output_path::String = joinpath(@__
     save(joinpath(output_path, "mimipagemontecarlooutput.csv"),df)
 end
 
-# function do_monte_carlo_runs(samplesize::Int, output_path::String = joinpath(@__DIR__, "../output"); scenario::String = "RCP4.5 & SSP2", )
-#     # get simulation
-#     mcs = getsim()
-#
-#     println(typeof(mcs))
-#
-#     # get a model
-#     m = getpage(scenario)
-#     run(m)
-#
-#     # continue Setup MC simulation
-#     set_models!(mcs, [m])
-#     generate_trials!(mcs, samplesize, filename = joinpath(output_path, "trialdata.csv"))
-#     #
-#     # Run
-#     res = run(mcs, m, samplesize; results_output_dir = output_path)
-#
-#     # reformat outputs for testing and analysis
-#     reformat_RV_outputs(samplesize, output_path=output_path)
-# end
 
-function do_monte_carlo_runs(samplesize::Int, scenario::String = "RCP4.5 & SSP2", output_path::String = joinpath(@__DIR__, "../output"))
+
+function do_monte_carlo_runs(samplesize::Int, scenario::String = "RCP4.5 & SSP2", output_path::String = joinpath(@__DIR__, "../../output"))
     # get simulation
     mcs = getsim()
 
@@ -362,7 +343,7 @@ function compute_scc_mcs(m::Model, samplesize::Int; year::Union{Int, Nothing} = 
     return scc_results
 end
 
-function get_scc_mcs(samplesize::Int, year::Int, output_path::String = joinpath(@__DIR__, "../output");
+function get_scc_mcs(samplesize::Int, year::Int, output_path::String = joinpath(@__DIR__, "../../output");
                       eta::Union{Float64, Nothing} = nothing, prtp::Union{Float64, Nothing} = nothing,
                       pulse_size::Union{Float64, Nothing} = 75000.,
                       scenario::String = "RCP4.5 & SSP2",
@@ -408,8 +389,3 @@ function get_scc_mcs(samplesize::Int, year::Int, output_path::String = joinpath(
 
     scc_results
 end
-
-# include("mcs.jl")
-# do_monte_carlo_runs(100)
-# include("compute_scc.jl")
-# get_scc_mcs(100, 2020)
