+++ src/extensions/runmodel_annual.jl
@@ -26,6 +26,9 @@ include("main_model_annual.jl")
 include("mcs_annual.jl")
 include("compute_scc_annual.jl")
 
+if !@isdefined samplesize
+    samplesize = 50000
+end
 
 for scenario in ["1.5 degC Target", "RCP2.6 & SSP1", "RCP4.5 & SSP2", "RCP8.5 & SSP5"]
     model = "PAGE-ANN"
@@ -37,15 +40,14 @@ for scenario in ["1.5 degC Target", "RCP2.6 & SSP1", "RCP4.5 & SSP2", "RCP8.5 & 
     # open up Explorer UI, for visual exploration of the variables
     # explore(m)
 
-    samplesize = 50000
     # do general monte carlo simulation and save the output
-    do_monte_carlo_runs(samplesize, scenario, joinpath(@__DIR__, "../output", scenario, model))
+    do_monte_carlo_runs(samplesize, scenario, joinpath(@__DIR__, "../../output", scenario, model))
 
     # get the social cost of carbon for the Monte Carlo simulations and save the output
     sccs = compute_scc_mcs(m, samplesize, year=2020)
     # store results in DataFrame
     df = DataFrame(Any[fill(model, samplesize), fill(scenario, samplesize), sccs], [:ModelName, :ScenarioName, :SCC])
     # write out to csv
-    DIR = joinpath(@__DIR__, "..", "output")
+    DIR = joinpath(@__DIR__, "..", "../output")
     CSV.write(joinpath(DIR, string(model, "_", scenario, "_", "scc.csv")), df)
 end
